pyproject.toml
[project]
name = "efflux-desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "anthropic>=0.49.0",
    "colorlog>=6.9.0",
    "fastapi>=0.115.12",
    "injector>=0.22.0",
    "mcp>=1.6.0",
    "openai>=1.72.0",
    "pdfminer-six>=20250327",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "socksio>=1.0.0",
    "uvicorn>=0.34.0",
]
